# frontend/public/openapi.json
{
  "components": {
    "schemas": {
...
      "Bill": {
        "properties": {
          "boundShipmentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "closedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "consignee": {
            "description": "0x0000… for a bill made out to order",
            "type": "string"
          },
          "documentHash": {
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/BillTransfer"
            },
            "type": "array"
          },
          "holder": {
            "type": "string"
          },
          "issuedAt": {
            "format": "date-time",
            "type": "string"
          },
          "issuer": {
            "type": "string"
          },
          "shipper": {
            "type": "string"
          },
          "status": {
            "enum": [
              "ISSUED",
              "SURRENDERED",
              "VOID"
            ],
            "type": "string"
          },
          "tokenId": {
            "type": "string"
          },
          "transfers": {
            "description": "endorsements after issue",
            "type": "integer"
          }
        },
        "required": [
          "boundShipmentId",
          "closedAt",
          "consignee",
          "documentHash",
          "history",
          "holder",
          "issuedAt",
          "issuer",
          "shipper",
          "status",
          "tokenId",
          "transfers"
        ],
        "type": "object"
      },
      "BillList": {
        "properties": {
          "bills": {
            "items": {
              "$ref": "#/components/schemas/Bill"
            },
            "type": "array"
          }
        },
        "required": [
          "bills"
        ],
        "type": "object"
      },
      "BillTransfer": {
        "properties": {
          "at": {
            "format": "date-time",
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "txHash": {
            "type": "string"
          }
        },
        "required": [
          "at",
          "from",
          "to",
          "txHash"
        ],
        "type": "object"
      },
...
      "Document": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "keccak256": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "matchesBill": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "for a bill_of_lading: the token id of the issued v3 eBL whose documentHash equals this keccak256, else null"
          },
          "matchesInvoiceHash": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "signer": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          }
        },
        "required": [
          "createdAt",
          "id",
          "keccak256",
          "kind",
          "matchesBill",
          "matchesInvoiceHash",
          "name",
          "role",
          "sha256",
          "signer",
          "sizeBytes"
        ],
        "type": "object"
      },
      "DocumentList": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array"
          }
        },
        "required": [
          "documents"
        ],
        "type": "object"
      },
      "DocumentRequest": {
        "properties": {
          "issuedAt": {
            "type": "integer"
          },
          "keccak256": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          }
        },
        "required": [
          "issuedAt",
          "keccak256",
          "kind",
          "name",
          "sha256",
          "signature",
          "sizeBytes"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "stable machine-readable code, e.g. invalid_request, unauthorized, replayed, chain_rejected",
                "type": "string"
              },
              "message": {
                "description": "human-readable explanation; never contains internal detail",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminKey": {
        "description": "the operator's ADMIN_API_KEY",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "alchemySignature": {
        "description": "Alchemy Notify: hex HMAC-SHA256 of the raw request body with the webhook's signing key (ALCHEMY_WEBHOOK_SIGNING_KEYS)",
        "in": "header",
        "name": "X-Alchemy-Signature",
        "type": "apiKey"
      },
      "sourceSignature": {
        "description": "Signed by a registered evidence source. The signing string is\n\n```\nCARGOFLOW-V1\\n<METHOD>\\n<PATH>\\n<TIMESTAMP>\\n<hex sha256(body)>\n```\n\n`ed25519` sources sign it with Ed25519. `p256` sources sign it with ECDSA P-256 over SHA-256 (X-Signature is the DER or raw 64-byte r||s signature). `webauthn` sources run navigator.credentials.get with challenge = sha256(signing string) (32 raw bytes) and send the assertion: X-Signature = signature, X-WebAuthn-Authenticator-Data = authenticatorData, X-WebAuthn-Client-Data = clientDataJSON (all base64url); the backend checks type `webauthn.get`, the challenge, the registered RP id hash, the user-present flag and the signature over authenticatorData || sha256(clientDataJSON).",
        "in": "header",
        "name": "X-Signature",
        "type": "apiKey"
      },
      "walletSignature": {
        "description": "Not a header: the body carries `issuedAt` and `signature` (hex EIP-191 personal_sign) over the operation's `x-cargoflow-signed-message`. Declared as a scheme so clients can see which operations need a wallet.",
        "in": "header",
        "name": "X-Wallet-Signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CargoFlow turns signed shipment telemetry into on-chain financing decisions. Public reads need no credentials. Writes are authenticated in one of three ways: the operator's admin key, a party's wallet signature over a fixed message (see each operation's `x-cargoflow-signed-message`), or an evidence source's request signature (Ed25519, P-256 or WebAuthn). Errors are `{\"error\": {\"code\", \"message\"}}` with stable codes. Amounts are USDG base units (6 decimals) as decimal strings.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "CargoFlow API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ebl": {
      "get": {
        "description": "404 on a deployment without the EBLRegistry. Designed around MLETR concepts (exclusive control, singularity, integrity); not a legal compliance claim.",
        "operationId": "listBills",
        "parameters": [
          {
            "description": "only bills this 0x address holds",
            "in": "query",
            "name": "holder",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillList"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "Electronic bills of lading (v3 EBLRegistry)",
        "tags": [
          "Standards"
        ]
      }
    },
    "/v1/ebl/{tokenId}": {
      "get": {
        "operationId": "getBill",
        "parameters": [
          {
            "description": "the bill's ERC-721 token id (decimal)",
            "in": "path",
            "name": "tokenId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "One bill of lading: parties, holder, status, transfer history and bound shipment",
        "tags": [
          "Standards"
        ]
      }
    },
...
    "/v1/shipments/{id}/documents": {
      "get": {
        "operationId": "listDocuments",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "Document attestations",
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "description": "Signed by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow document\nshipment: <id>\nkind: <kind>\nsha256: <0x..>\nissued: <t>\n```",
        "operationId": "attestDocument",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "the signer attested this file before"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Attest a file by its hashes (a party)",
        "tags": [
          "Documents"
        ],
        "x-cargoflow-signed-message": "CargoFlow document\nshipment: <id>\nkind: <kind>\nsha256: <0x..>\nissued: <t>"
      }
    },
...
  },
  "servers": [
    {
      "description": "this backend",
      "url": "/"
    }
  ]
}